+++ src/agent/agreement.rs
@@ -5,9 +5,11 @@ use crate::agent::Agent;
 use crate::agent::JACS_VERSION_DATE_FIELDNAME;
 use crate::agent::JACS_VERSION_FIELDNAME;
 use crate::agent::{
-    AGENT_AGREEMENT_FIELDNAME, DOCUMENT_AGREEMENT_HASH_FIELDNAME, JACS_PREVIOUS_VERSION_FIELDNAME,
-    SHA256_FIELDNAME,
+    AGENT_AGREEMENT_FIELDNAME, DOCUMENT_AGREEMENT_HASH_FIELDNAME, JACS_IGNORE_FIELDS,
+    JACS_PREVIOUS_VERSION_FIELDNAME, SHA256_FIELDNAME, TASK_END_AGREEMENT_FIELDNAME,
+    TASK_START_AGREEMENT_FIELDNAME,
 };
+
 use crate::crypt::hash::hash_public_key;
 use crate::crypt::hash::hash_string;
 use crate::schema::utils::ValueExt;
@@ -29,52 +31,76 @@ pub trait Agreement {
         agentids: &Vec<String>,
         question: Option<&String>,
         context: Option<&String>,
+        agreement_fieldname: Option<String>,
     ) -> Result<JACSDocument, Box<dyn Error>>;
     /// given a document id and a list of agents, return an updated document
     fn add_agents_to_agreement(
         &mut self,
         document_key: &String,
         agentids: &Vec<String>,
+        agreement_fieldname: Option<String>,
     ) -> Result<JACSDocument, Box<dyn Error>>;
     /// given a document id and a list of agents, return an updated document
     fn remove_agents_from_agreement(
         &mut self,
         document_key: &String,
         agentids: &Vec<String>,
+        agreement_fieldname: Option<String>,
     ) -> Result<JACSDocument, Box<dyn Error>>;
     /// given a document id sign a document, return an updated document
-    fn sign_agreement(&mut self, document_key: &String) -> Result<JACSDocument, Box<dyn Error>>;
+    fn sign_agreement(
+        &mut self,
+        document_key: &String,
+        agreement_fieldname: Option<String>,
+    ) -> Result<JACSDocument, Box<dyn Error>>;
     /// given a document, check all agreement signatures
-    fn check_agreement(&self, document_key: &String) -> Result<String, Box<dyn Error>>;
+    fn check_agreement(
+        &self,
+        document_key: &String,
+        agreement_fieldname: Option<String>,
+    ) -> Result<String, Box<dyn Error>>;
 
     /// agreements update documents
     /// however this updates the document, which updates, version, lastversion and version date
     /// the agreement itself needs it's own hash to track
     /// on a subset of fields
     /// the standard hash detects any changes at all
-    fn agreement_hash(&self, value: Value) -> Result<String, Box<dyn Error>>;
+    fn agreement_hash(
+        &self,
+        value: Value,
+        agreement_fieldname: &String,
+    ) -> Result<String, Box<dyn Error>>;
 
     /// remove fields that should not be used for agreement signature
     fn trim_fields_for_hashing_and_signing(
         &self,
         value: Value,
+        agreement_fieldname: &String,
     ) -> Result<(String, Vec<String>), Box<dyn Error>>;
 
     fn agreement_get_question_and_context(
         &self,
         document_key: &std::string::String,
+        agreement_fieldname: Option<String>,
     ) -> Result<(String, String), Box<dyn Error>>;
 }
 
 impl Agreement for Agent {
-    fn agreement_hash(&self, value: Value) -> Result<String, Box<dyn Error>> {
-        let (values_as_string, _fields) = self.trim_fields_for_hashing_and_signing(value)?;
+    fn agreement_hash(
+        &self,
+        value: Value,
+        agreement_fieldname: &String,
+    ) -> Result<String, Box<dyn Error>> {
+        let (values_as_string, _fields) =
+            self.trim_fields_for_hashing_and_signing(value, agreement_fieldname)?;
         Ok(hash_string(&values_as_string))
     }
 
+    // ignore these extra fields will change
     fn trim_fields_for_hashing_and_signing(
         &self,
         value: Value,
+        agreement_fieldname: &String,
     ) -> Result<(String, Vec<String>), Box<dyn Error>> {
         let mut new_obj: Value = value.clone();
         new_obj.as_object_mut().map(|obj| {
@@ -85,7 +111,7 @@ impl Agreement for Agent {
         });
 
         let (values_as_string, fields) =
-            Agent::get_values_as_string(&new_obj, None, &AGENT_AGREEMENT_FIELDNAME.to_string())?;
+            Agent::get_values_as_string(&new_obj, None, &agreement_fieldname)?;
         return Ok((values_as_string, fields));
     }
 
@@ -95,7 +121,12 @@ impl Agreement for Agent {
         agentids: &Vec<String>,
         question: Option<&String>,
         context: Option<&String>,
+        agreement_fieldname: Option<String>,
     ) -> Result<JACSDocument, Box<(dyn StdError + 'static)>> {
+        let agreement_fieldname_key = match agreement_fieldname {
+            Some(key) => key,
+            _ => AGENT_AGREEMENT_FIELDNAME.to_string(),
+        };
         let document = self.get_document(document_key)?;
         let mut value = document.value;
 
@@ -109,9 +140,10 @@ impl Agreement for Agent {
             _ => "",
         };
         // todo error if value[AGENT_AGREEMENT_FIELDNAME] exists.validate
-        let agreement_hash_value = json!(self.agreement_hash(value.clone())?);
+        let agreement_hash_value =
+            json!(self.agreement_hash(value.clone(), &agreement_fieldname_key)?);
         value[DOCUMENT_AGREEMENT_HASH_FIELDNAME] = agreement_hash_value.clone();
-        value[AGENT_AGREEMENT_FIELDNAME] = json!({
+        value[agreement_fieldname_key.clone()] = json!({
             // based on v1
             "signatures": [],
             "agentIDs": agentids,
@@ -122,7 +154,7 @@ impl Agreement for Agent {
             self.update_document(document_key, &serde_json::to_string(&value)?, None, None)?;
 
         let agreement_hash_value_after =
-            json!(self.agreement_hash(updated_document.value.clone())?);
+            json!(self.agreement_hash(updated_document.value.clone(), &agreement_fieldname_key)?);
         // could be unit test, but want this in for safety
         if agreement_hash_value != agreement_hash_value_after {
             return Err(format!(
@@ -144,12 +176,17 @@ impl Agreement for Agent {
         &mut self,
         document_key: &std::string::String,
         agentids: &Vec<String>,
+        agreement_fieldname: Option<String>,
     ) -> Result<JACSDocument, Box<(dyn StdError + 'static)>> {
+        let agreement_fieldname_key = match agreement_fieldname {
+            Some(key) => key,
+            _ => AGENT_AGREEMENT_FIELDNAME.to_string(),
+        };
         let document = self.get_document(document_key)?;
         let mut value = document.value;
         let binding = value[DOCUMENT_AGREEMENT_HASH_FIELDNAME].clone();
 
-        if let Some(jacs_agreement) = value.get_mut(AGENT_AGREEMENT_FIELDNAME) {
+        if let Some(jacs_agreement) = value.get_mut(agreement_fieldname_key) {
             if let Some(agents) = jacs_agreement.get_mut("agentIDs") {
                 if let Some(agents_array) = agents.as_array_mut() {
                     let merged_agents = subtract_vecs(
@@ -180,12 +217,17 @@ impl Agreement for Agent {
         &mut self,
         document_key: &std::string::String,
         agentids: &Vec<String>,
+        agreement_fieldname: Option<String>,
     ) -> Result<JACSDocument, Box<(dyn StdError + 'static)>> {
+        let agreement_fieldname_key = match agreement_fieldname {
+            Some(key) => key,
+            _ => AGENT_AGREEMENT_FIELDNAME.to_string(),
+        };
         let document = self.get_document(document_key)?;
         let mut value = document.value;
         let binding = value[DOCUMENT_AGREEMENT_HASH_FIELDNAME].clone();
 
-        if let Some(jacs_agreement) = value.get_mut(AGENT_AGREEMENT_FIELDNAME) {
+        if let Some(jacs_agreement) = value.get_mut(agreement_fieldname_key.clone()) {
             if let Some(agents) = jacs_agreement.get_mut("agentIDs") {
                 if let Some(agents_array) = agents.as_array_mut() {
                     let merged_agents = merge_without_duplicates(
@@ -203,7 +245,7 @@ impl Agreement for Agent {
                 jacs_agreement["agentIDs"] = json!(agentids);
             }
         } else {
-            value[AGENT_AGREEMENT_FIELDNAME] = json!({
+            value[agreement_fieldname_key] = json!({
                 "agentIDs": agentids,
                 "signatures": [],
             });
@@ -220,26 +262,36 @@ impl Agreement for Agent {
     fn sign_agreement(
         &mut self,
         document_key: &std::string::String,
-    ) -> Result<JACSDocument, Box<(dyn StdError + 'static)>> {
+        agreement_fieldname: Option<String>,
+    ) -> Result<JACSDocument, Box<dyn Error>> {
+        let agreement_fieldname_key = match agreement_fieldname {
+            Some(ref key) => key.to_string(),
+            _ => AGENT_AGREEMENT_FIELDNAME.to_string(),
+        };
+
         let document = self.get_document(document_key)?;
         let mut value = document.value;
         let binding = value[DOCUMENT_AGREEMENT_HASH_FIELDNAME].clone();
         let original_agreement_hash_value = binding.as_str();
         // todo use this
-        let _calculated_agreement_hash_value = self.agreement_hash(value.clone())?;
+        let _calculated_agreement_hash_value =
+            self.agreement_hash(value.clone(), &agreement_fieldname_key)?;
         let signing_agent_id = self.get_id().expect("agent id");
         //  generate signature object
         let (_values_as_string, fields) =
-            self.trim_fields_for_hashing_and_signing(value.clone())?;
+            self.trim_fields_for_hashing_and_signing(value.clone(), &agreement_fieldname_key)?;
         let agents_signature: Value = self.signing_procedure(
             &value.clone(),
             Some(&fields),
-            &AGENT_AGREEMENT_FIELDNAME.to_string(),
+            &agreement_fieldname_key.to_string(),
         )?;
 
         // redundant but make sure agent is listed as a signatory
-        let agent_complete_document =
-            self.add_agents_to_agreement(document_key, &vec![signing_agent_id.clone()])?;
+        let agent_complete_document = self.add_agents_to_agreement(
+            document_key,
+            &vec![signing_agent_id.clone()],
+            agreement_fieldname,
+        )?;
         value = agent_complete_document.getvalue();
         let agent_complete_key = agent_complete_document.getkey();
         debug!(
@@ -247,7 +299,7 @@ impl Agreement for Agent {
             serde_json::to_string_pretty(&agents_signature).expect("agents_signature print")
         );
 
-        if let Some(jacs_agreement) = value.get_mut(AGENT_AGREEMENT_FIELDNAME) {
+        if let Some(jacs_agreement) = value.get_mut(&agreement_fieldname_key) {
             if let Some(signatures) = jacs_agreement.get_mut("signatures") {
                 if let Some(signatures_array) = signatures.as_array_mut() {
                     signatures_array.push(agents_signature);
@@ -258,7 +310,7 @@ impl Agreement for Agent {
                 jacs_agreement["signatures"] = json!([agents_signature]);
             }
         } else {
-            value[AGENT_AGREEMENT_FIELDNAME] = json!({
+            value[agreement_fieldname_key.clone()] = json!({
                 "agentIDs": [signing_agent_id],
                 "signatures": [agents_signature]
             });
@@ -271,7 +323,8 @@ impl Agreement for Agent {
             None,
         )?;
 
-        let agreement_hash_value_after = self.agreement_hash(updated_document.value.clone())?;
+        let agreement_hash_value_after =
+            self.agreement_hash(updated_document.value.clone(), &agreement_fieldname_key)?;
 
         // could be unit test, but want this in for safety
         if original_agreement_hash_value != Some(&agreement_hash_value_after) {
@@ -293,18 +346,25 @@ impl Agreement for Agent {
     fn agreement_get_question_and_context(
         &self,
         document_key: &std::string::String,
+        agreement_fieldname: Option<String>,
     ) -> Result<(String, String), Box<dyn Error>> {
+        let agreement_fieldname_key = match agreement_fieldname {
+            Some(key) => key,
+            _ => AGENT_AGREEMENT_FIELDNAME.to_string(),
+        };
+
         let document = self.get_document(document_key)?;
         let error_message = format!("{} missing", DOCUMENT_AGREEMENT_HASH_FIELDNAME);
         let original_agreement_hash_value = document.value[DOCUMENT_AGREEMENT_HASH_FIELDNAME]
             .as_str()
             .expect(&error_message);
-        let calculated_agreement_hash_value = self.agreement_hash(document.value.clone())?;
+        let calculated_agreement_hash_value =
+            self.agreement_hash(document.value.clone(), &agreement_fieldname_key)?;
         if original_agreement_hash_value != calculated_agreement_hash_value {
             return Err("check_agreement: agreement hashes don't match".into());
         }
 
-        if let Some(jacs_agreement) = document.value.get(AGENT_AGREEMENT_FIELDNAME) {
+        if let Some(jacs_agreement) = document.value.get(agreement_fieldname_key) {
             let question = jacs_agreement
                 .get_str("question")
                 .expect("agreement_get_question_and_context question field");
@@ -322,29 +382,36 @@ impl Agreement for Agent {
     fn check_agreement(
         &self,
         document_key: &std::string::String,
+        agreement_fieldname: Option<String>,
     ) -> Result<String, Box<(dyn StdError + 'static)>> {
+        let agreement_fieldname_key: String = match agreement_fieldname {
+            Some(ref key) => key.to_string(),
+            _ => AGENT_AGREEMENT_FIELDNAME.to_string(),
+        };
+
         let document = self.get_document(document_key)?;
         let local_doc_value = document.value.clone();
         let error_message = format!("{} missing", DOCUMENT_AGREEMENT_HASH_FIELDNAME);
         let original_agreement_hash_value = document.value[DOCUMENT_AGREEMENT_HASH_FIELDNAME]
             .as_str()
             .expect(&error_message);
-        let calculated_agreement_hash_value = self.agreement_hash(document.value.clone())?;
+        let calculated_agreement_hash_value =
+            self.agreement_hash(document.value.clone(), &agreement_fieldname_key)?;
         if original_agreement_hash_value != calculated_agreement_hash_value {
             return Err("check_agreement: agreement hashes don't match".into());
         }
 
-        let unsigned = document.agreement_unsigned_agents()?;
+        let unsigned = document.agreement_unsigned_agents(agreement_fieldname.clone())?;
         if unsigned.len() > 0 {
             return Err(format!(
                 "not all agents have signed: {:?} {:?}",
                 unsigned,
-                document.value.get(AGENT_AGREEMENT_FIELDNAME).unwrap()
+                document.value.get(agreement_fieldname_key).unwrap()
             )
             .into());
         }
 
-        if let Some(jacs_agreement) = document.value.get(AGENT_AGREEMENT_FIELDNAME) {
+        if let Some(jacs_agreement) = document.value.get(agreement_fieldname_key.clone()) {
             if let Some(signatures) = jacs_agreement.get("signatures") {
                 if let Some(signatures_array) = signatures.as_array() {
                     for signature in signatures_array {
@@ -388,12 +455,15 @@ impl Agreement for Agent {
                             "testing agreement sig agent_id_and_version {} {} {} ",
                             agent_id_and_version, noted_hash, public_key_enc_type
                         );
-                        let (_values_as_string, fields) =
-                            self.trim_fields_for_hashing_and_signing(local_doc_value.clone())?;
+                        let (_values_as_string, fields) = self
+                            .trim_fields_for_hashing_and_signing(
+                                local_doc_value.clone(),
+                                &agreement_fieldname_key,
+                            )?;
                         let result = self.signature_verification_procedure(
                             &document.value,
                             Some(&fields),
-                            &AGENT_AGREEMENT_FIELDNAME.to_string(),
+                            &agreement_fieldname_key.to_string(),
                             agents_public_key,
                             Some(public_key_enc_type.clone()),
                             Some(noted_hash.clone()),
@@ -421,7 +491,7 @@ pub fn merge_without_duplicates(vec1: &Vec<String>, vec2: &Vec<String>) -> Vec<S
 }
 
 pub fn subtract_vecs(vec1: &Vec<String>, vec2: &Vec<String>) -> Vec<String> {
-    println!("subtract_vecs A {:?} {:?} ", vec1, vec2);
+    debug!("subtract_vecs A {:?} {:?} ", vec1, vec2);
 
     let to_remove: HashSet<&String> = vec2.iter().collect();
     let return_vec1 = vec1
@@ -429,6 +499,6 @@ pub fn subtract_vecs(vec1: &Vec<String>, vec2: &Vec<String>) -> Vec<String> {
         .filter(|item| !to_remove.contains(item))
         .cloned()
         .collect();
-    println!("subtract_vecs B {:?}- {:?} = {:?}", vec1, vec2, return_vec1);
+    debug!("subtract_vecs B {:?}- {:?} = {:?}", vec1, vec2, return_vec1);
     return return_vec1;
 }