+++ src/agent/document.rs
@@ -13,6 +13,7 @@ use chrono::Utc;
 use difference::{Changeset, Difference};
 use flate2::read::GzDecoder;
 use log::error;
+use regex::Regex;
 use serde_json::json;
 use serde_json::Value;
 use sha2::{Digest, Sha256};
@@ -44,9 +45,34 @@ impl JACSDocument {
         self.value.clone()
     }
 
-    pub fn agreement_unsigned_agents(&self) -> Result<Vec<String>, Box<dyn Error>> {
-        let all_requested_agents = self.agreement_requested_agents()?;
-        let all_agreement_signed_agents = self.agreement_signed_agents()?;
+    pub fn getschema(&self) -> Result<String, Box<dyn Error>> {
+        let schemafield = "$schema";
+        if let Some(schema) = self.value.get(schemafield) {
+            if let Some(schema_str) = schema.as_str() {
+                return Ok(schema_str.to_string());
+            }
+        }
+        return Err("no schema in doc or schema is not a string".into());
+    }
+
+    pub fn getshortschema(&self) -> Result<String, Box<dyn Error>> {
+        let longschema = self.getschema()?;
+        let re = Regex::new(r"/([^/]+)\.schema\.json$").unwrap();
+
+        if let Some(caps) = re.captures(&longschema) {
+            if let Some(matched) = caps.get(1) {
+                return Ok(matched.as_str().to_string());
+            }
+        }
+        Err("Failed to extract schema name from URL".into())
+    }
+
+    pub fn agreement_unsigned_agents(
+        &self,
+        agreement_fieldname: Option<String>,
+    ) -> Result<Vec<String>, Box<dyn Error>> {
+        let all_requested_agents = self.agreement_requested_agents(agreement_fieldname.clone())?;
+        let all_agreement_signed_agents = self.agreement_signed_agents(agreement_fieldname)?;
 
         return Ok(subtract_vecs(
             &all_requested_agents,
@@ -54,9 +80,16 @@ impl JACSDocument {
         ));
     }
 
-    pub fn agreement_requested_agents(&self) -> Result<Vec<String>, Box<dyn Error>> {
+    pub fn agreement_requested_agents(
+        &self,
+        agreement_fieldname: Option<String>,
+    ) -> Result<Vec<String>, Box<dyn Error>> {
+        let agreement_fieldname_key = match agreement_fieldname {
+            Some(key) => key,
+            _ => AGENT_AGREEMENT_FIELDNAME.to_string(),
+        };
         let value: &serde_json::Value = &self.value;
-        if let Some(jacs_agreement) = value.get(AGENT_AGREEMENT_FIELDNAME) {
+        if let Some(jacs_agreement) = value.get(agreement_fieldname_key) {
             if let Some(agents) = jacs_agreement.get("agentIDs") {
                 if let Some(agents_array) = agents.as_array() {
                     return Ok(agents_array
@@ -77,9 +110,16 @@ impl JACSDocument {
         return Err("no agreement or signatures in agreement".into());
     }
 
-    pub fn agreement_signed_agents(&self) -> Result<Vec<String>, Box<dyn Error>> {
+    pub fn agreement_signed_agents(
+        &self,
+        agreement_fieldname: Option<String>,
+    ) -> Result<Vec<String>, Box<dyn Error>> {
+        let agreement_fieldname_key = match agreement_fieldname {
+            Some(key) => key,
+            _ => AGENT_AGREEMENT_FIELDNAME.to_string(),
+        };
         let value: &serde_json::Value = &self.value;
-        if let Some(jacs_agreement) = value.get(AGENT_AGREEMENT_FIELDNAME) {
+        if let Some(jacs_agreement) = value.get(agreement_fieldname_key) {
             if let Some(signatures) = jacs_agreement.get("signatures") {
                 if let Some(signatures_array) = signatures.as_array() {
                     let mut signed_agents: Vec<String> = Vec::<String>::new();