+++ tests/agreement_test.rs
@@ -2,6 +2,7 @@ use jacs::agent::agreement::Agreement;
 use jacs::agent::boilerplate::BoilerPlate;
 use jacs::agent::document::Document;
 use jacs::agent::loaders::FileLoader;
+use jacs::agent::AGENT_AGREEMENT_FIELDNAME;
 use jacs::crypt::KeyManager;
 use secrecy::ExposeSecret;
 mod utils;
@@ -26,13 +27,25 @@ fn test_create_agreement() {
     let document_key = document.getkey();
     // agent one creates agreement document
     let unsigned_doc = agent
-        .create_agreement(&document_key, &agentids, None, None)
+        .create_agreement(
+            &document_key,
+            &agentids,
+            None,
+            None,
+            Some(AGENT_AGREEMENT_FIELDNAME.to_string()),
+        )
         .expect("create_agreement");
 
     println!("{}", unsigned_doc.to_string());
 
     // agent one  tries and fails to creates agreement document
-    let _result = agent.create_agreement(&document_key, &agentids, None, None);
+    let _result = agent.create_agreement(
+        &document_key,
+        &agentids,
+        None,
+        None,
+        Some(AGENT_AGREEMENT_FIELDNAME.to_string()),
+    );
 
     // agent two signs document
 
@@ -58,36 +71,61 @@ fn test_add_and_remove_agents() {
     let document = agent.load_document(&document_string).unwrap();
     let document_key = document.getkey();
     let mut doc_v1 = agent
-        .create_agreement(&document_key, &agents_orig, None, None)
+        .create_agreement(
+            &document_key,
+            &agents_orig,
+            None,
+            None,
+            Some(AGENT_AGREEMENT_FIELDNAME.to_string()),
+        )
         .expect("create_agreement");
     let doc_v1_key = doc_v1.getkey();
     println!(
         "doc_v1_key agents requested {:?} unsigned {:?}",
-        doc_v1.agreement_requested_agents().unwrap(),
-        doc_v1.agreement_unsigned_agents().unwrap()
+        doc_v1
+            .agreement_requested_agents(Some(AGENT_AGREEMENT_FIELDNAME.to_string()),)
+            .unwrap(),
+        doc_v1
+            .agreement_unsigned_agents(Some(AGENT_AGREEMENT_FIELDNAME.to_string()),)
+            .unwrap()
     );
     let mut doc_v2 = agent
-        .add_agents_to_agreement(&doc_v1_key, &agents_to_add)
+        .add_agents_to_agreement(
+            &doc_v1_key,
+            &agents_to_add,
+            Some(AGENT_AGREEMENT_FIELDNAME.to_string()),
+        )
         .expect("add_agents_to_agreement");
     let doc_v2_key = doc_v2.getkey();
     println!(
         "doc_v2_key agents {:?}",
-        doc_v2.agreement_requested_agents().unwrap()
+        doc_v2
+            .agreement_requested_agents(Some(AGENT_AGREEMENT_FIELDNAME.to_string()),)
+            .unwrap()
     );
     let mut doc_v3 = agent
-        .remove_agents_from_agreement(&doc_v2_key, &agents_to_remove)
+        .remove_agents_from_agreement(
+            &doc_v2_key,
+            &agents_to_remove,
+            Some(AGENT_AGREEMENT_FIELDNAME.to_string()),
+        )
         .expect("remove_agents_from_agreement");
     let doc_v3_key = doc_v3.getkey();
     println!(
         "doc_v3 agents {:?}",
-        doc_v3.agreement_requested_agents().unwrap()
+        doc_v3
+            .agreement_requested_agents(Some(AGENT_AGREEMENT_FIELDNAME.to_string()),)
+            .unwrap()
     );
 
     // println!(
     //     "final signature requests were\n {}",
     //     serde_json::to_string_pretty(&doc_v3.value).expect("pretty print")
     // );
-    let result = agent.check_agreement(&doc_v3.getkey());
+    let result = agent.check_agreement(
+        &doc_v3.getkey(),
+        Some(AGENT_AGREEMENT_FIELDNAME.to_string()),
+    );
     match result {
         Err(_) => assert!(true),
         Ok(_) => assert!(false),
@@ -126,13 +164,22 @@ fn test_sign_agreement() {
     let document_key = document.getkey();
     // agent one creates agreement document
     let unsigned_doc = agent
-        .create_agreement(&document_key, &agentids, None, None)
+        .create_agreement(
+            &document_key,
+            &agentids,
+            None,
+            None,
+            Some(AGENT_AGREEMENT_FIELDNAME.to_string()),
+        )
         .expect("create_agreement");
 
     let unsigned_doc_key = unsigned_doc.getkey();
 
     let signed_document = agent
-        .sign_agreement(&unsigned_doc_key)
+        .sign_agreement(
+            &unsigned_doc_key,
+            Some(AGENT_AGREEMENT_FIELDNAME.to_string()),
+        )
         .expect("signed_document ");
     let signed_document_key = signed_document.getkey();
     let signed_document_string =
@@ -140,7 +187,10 @@ fn test_sign_agreement() {
 
     let _ = agent_two.load_document(&signed_document_string).unwrap();
     let both_signed_document = agent_two
-        .sign_agreement(&signed_document_key)
+        .sign_agreement(
+            &signed_document_key,
+            Some(AGENT_AGREEMENT_FIELDNAME.to_string()),
+        )
         .expect("signed_document ");
 
     println!(
@@ -153,12 +203,21 @@ fn test_sign_agreement() {
 
     println!(
         "both_signed_document agents requested {:?} unsigned {:?} signed {:?}",
-        both_signed_document.agreement_requested_agents().unwrap(),
-        both_signed_document.agreement_unsigned_agents().unwrap(),
-        both_signed_document.agreement_signed_agents().unwrap()
+        both_signed_document
+            .agreement_requested_agents(Some(AGENT_AGREEMENT_FIELDNAME.to_string()))
+            .unwrap(),
+        both_signed_document
+            .agreement_unsigned_agents(Some(AGENT_AGREEMENT_FIELDNAME.to_string()),)
+            .unwrap(),
+        both_signed_document
+            .agreement_signed_agents(Some(AGENT_AGREEMENT_FIELDNAME.to_string()),)
+            .unwrap()
     );
 
-    let result = agent_two.check_agreement(&both_signed_document.getkey());
+    let result = agent_two.check_agreement(
+        &both_signed_document.getkey(),
+        Some(AGENT_AGREEMENT_FIELDNAME.to_string()),
+    );
     match result {
         Err(err) => {
             println!("{}", err);
@@ -172,7 +231,10 @@ fn test_sign_agreement() {
 
     let agent_one_both_signed_document = agent.load_document(&both_signed_document_string).unwrap();
     let agent_one_both_signed_document_key = agent_one_both_signed_document.getkey();
-    let result = agent.check_agreement(&agent_one_both_signed_document_key);
+    let result = agent.check_agreement(
+        &agent_one_both_signed_document_key,
+        Some(AGENT_AGREEMENT_FIELDNAME.to_string()),
+    );
     match result {
         Err(err) => {
             println!("{}", err);
@@ -181,7 +243,10 @@ fn test_sign_agreement() {
         Ok(_) => assert!(true),
     }
     let (question, context) = agent
-        .agreement_get_question_and_context(&agent_one_both_signed_document_key)
+        .agreement_get_question_and_context(
+            &agent_one_both_signed_document_key,
+            Some(AGENT_AGREEMENT_FIELDNAME.to_string()),
+        )
         .unwrap();
     println!(" question {}, context {}", question, context);
 }