+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 0.2.10
+
+- decouple message from task so they can arrive out of order. can be used to create context for task
+- parameteraize agreement field
+- task start and end agreements functions
+- fixed issue with schema path not being found, so list of fields returned incorrect
+- retrieve long and short schema name from docs - mostly for task and agent
+
+
 ## 0.2.9
 
 - tests for task and actions creation